# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.CreateOwnerRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone_number:
        type: string
    required:
    - email
    - name
    - password
    - phone_number
    type: object
  model.CreateOwnerResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      phone_number:
        type: string
    type: object
...
  model.GetOwnerResponse:
    properties:
      date_of_birth:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone_number:
        type: string
    type: object
...
  model.OwnerResponse:
    properties:
      owner:
        type: object
    type: object
  model.ResetPasswordRequest:
    properties:
      password:
        type: string
      password_confirm:
        type: string
    required:
    - password
    - password_confirm
    type: object
...
  model.UpdateOwnerRequest:
    properties:
      date_of_birth:
        type: string
      name:
        type: string
      phone_number:
        type: string
    required:
    - name
    type: object
  web.ErrJSONResponse:
    properties:
      error:
        $ref: '#/definitions/web.Error'
        type: object
      process_time:
        type: string
      status:
        type: string
      success:
        example: false
        type: boolean
    type: object
  web.Error:
    properties:
      message:
        type: string
    type: object
  web.JSONResponse:
    properties:
      data:
        type: object
        x-omitempty: true
      process_time:
        type: string
      status:
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:9000
info:
  contact:
    email: support.family-catering@example.com
    name: Family catering API Support
    url: http://www.family-catering.com/support
  description: Documentation for Family Catering API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Family Catering API
  version: "1.0"
paths:
...
  /owner:
    get:
      description: Show list of owners by (optionally) given limit and/or offset
      parameters:
      - description: Pagination limit
        format: int64
        in: query
        name: limit
        type: integer
      - description: Pagination offset
        format: int64
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            allOf:
            - $ref: '#/definitions/web.JSONResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/model.OwnerResponse'
                  - properties:
                      owner:
                        items:
                          $ref: '#/definitions/model.GetOwnerResponse'
                        type: array
                    type: object
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
      summary: Show list of owners
      tags:
      - owner
    post:
      consumes:
      - application/json
      description: Create an owner through Sign up
      parameters:
      - description: Create owner payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.CreateOwnerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            allOf:
            - $ref: '#/definitions/web.JSONResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/model.OwnerResponse'
                  - properties:
                      owner:
                        $ref: '#/definitions/model.CreateOwnerResponse'
                    type: object
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "409":
          description: Email already registered
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
      summary: Create an owner
      tags:
      - owner
  /owner/{id}:
    delete:
      description: Delete owner by given owner's id
      parameters:
      - description: Owner id
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - default: sid=<Add your session id>
        description: Session id
        in: header
        name: Cookie
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/web.JSONResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "404":
          description: Owner not found
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
      summary: Delete owner
      tags:
      - owner
    get:
      description: Show interest owner details
      parameters:
      - description: Owner id
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            allOf:
            - $ref: '#/definitions/web.JSONResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/model.OwnerResponse'
                  - properties:
                      owner:
                        $ref: '#/definitions/model.GetOwnerResponse'
                    type: object
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "404":
          description: Owner not found
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
      summary: Get owner by given id
      tags:
      - owner
    put:
      description: Update owner by given owner's id
      parameters:
      - description: Owner id
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Update owner payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.UpdateOwnerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/web.JSONResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "404":
          description: Owner not found
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
      summary: Update owner
      tags:
      - owner
  /owner/{id}/reset-password:
    put:
      description: Reset owner password by given id
      parameters:
      - description: Owner reset password id
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - default: sid=<Add your session id>
        description: Session id
        in: header
        name: Cookie
        required: true
        type: string
      - description: Reset Password payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/web.JSONResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "404":
          description: Owner not found
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
      summary: Reset owner password (logged in state)
      tags:
      - owner
  /owner/{id}/update-email:
    put:
      description: Update owner's email by given id
      parameters:
      - description: Owner reset password id
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - default: sid=<Add your session id>
        description: Session id
        in: header
        name: Cookie
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/web.JSONResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "404":
          description: Owner not found
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
      summary: Update Email owner
      tags:
      - owner
  /owner/reset-password/{rpid}:
    put:
      description: Reset owner password by given rpid
      parameters:
      - description: Owner reset password id
        in: path
        name: rpid
        required: true
        type: string
      - default: Bearer <Add password token here>
        description: Insert your password token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Reset Password payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/web.JSONResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "404":
          description: Owner not found
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
      summary: Reset owner password (logged out state)
      tags:
      - owner
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"